+++ src/cleansia_android/customer-app/src/main/java/cz/cleansia/customer/navigation/Routes.kt
@@ -18,6 +18,9 @@ object Routes {
     data object SignIn
 
     @Serializable
+    data object GuestOrder
+
+    @Serializable
     data object SignUp
 
     @Serializable